requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncmy]==2.0.23
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
asyncmy==0.2.9
cryptography==41.0.7
cachetools==5.3.2
aiosqlite==0.19.0

